[bugfix] Fix inconsistent assumptions about the size of ECLMaterialLawManager::materialLawParams_
In the functions ECLMaterialLawManager::materialLawParam(unsigned
elemetIdx) there is a switch on
ECLMaterialLawManager::hasElementSpecificParameters(). If true the
size is assumed to equal to the number elements in the grid, other it
is supposed to be the number of satNumRegions.

Note that this function is used when calculating the capillary
pressure in  ECLMaterialLawManager::swatInit.

Nevertheless, during ECLMaterialLawManager::init the same switch is
used to either call ECLMaterialLawManager::initElemSpecific_ or
ECLMaterialLawManager::initNonElemSpecific_. But both methods create
ECLMaterialLawManager::materialLawParams_ with the same size as the
number of elements and initialize one entry for each element.

This commit makes the assumption about the size of
ECLMaterialLawManager::materialLawParams_ in
ECLMaterialLawManager::materialLawParams() the same as in the init
method.