implement non-trivial capillar pressure hysteresis
So far, the imbibition curves were simply ignored. The reason why the
decks that we've tested so far work is that the capillary pressure
curves for imbibition are exactly the same as for drainage for these
decks.

Note: This code works even for capillary pressure curves for which the
endpoints do not match (as required by E100). This requires to use
linear extrapolation of the capillary pressure curves, though.

I've tested this on the Norne deck: Performance and results seem to be
unaffected.