Extrapolate for values not in domain (UniformTabulated2DFunction).
In the current code the interpolation actually already works if the
values are outside of the tabulated region.

With this change there is now a static variable
throwUntabulated (default: false) that we can set in opm-simulators.
If it is false we will interpolate for every value instead of throwing
and aborting (was the case always before).